Implement load product attributes on product detail UI

+++ src/Server/Core/Camino.Core/Contracts/Services/Products/IProductCategoryService.cs
@@ -12,10 +12,10 @@ namespace Camino.Core.Contracts.Services.Products
 {
     public interface IProductCategoryService
     {
-        Task<ProductCategoryResult> FindAsync(long id);
+        Task<ProductCategoryResult> FindAsync(int id);
         Task<BasePageList<ProductCategoryResult>> GetAsync(ProductCategoryFilter filter);
-        Task<IList<ProductCategoryResult>> SearchAsync(long[] currentIds, string search = "", int page = 1, int pageSize = 10);
-        Task<IList<ProductCategoryResult>> SearchParentsAsync(long[] currentIds, string search = "", int page = 1, int pageSize = 10);
+        Task<IList<ProductCategoryResult>> SearchAsync(int[] currentIds, string search = "", int page = 1, int pageSize = 10);
+        Task<IList<ProductCategoryResult>> SearchParentsAsync(int[] currentIds, string search = "", int page = 1, int pageSize = 10);
         List<ProductCategoryResult> Get(Expression<Func<ProductCategory, bool>> filter);
         Task<int> CreateAsync(ProductCategoryRequest request);
         Task<bool> UpdateAsync(ProductCategoryRequest request);
